requirements.txt
# This is an implicit value, here for clarity
--index-url https://pypi.python.org/simple/

jsonrpcserver==3.5.3
croniter==0.3.20
smbus-cffi==0.5.1
google-api-python-client==1.6.6
RPi.GPIO==0.6.3
requests==2.18.4
urllib3==1.22

